Show enclosing

+++ src/main/java/com/lp/drawingide/model/AbstractShape.java
@@ -13,22 +13,22 @@ public class AbstractShape {
         this.shape = shape;
         enclosing = new Enclosing();
         if (shape.getPoints().size() > 0) {
-            enclosing.setXMin(shape.getPoints().get(0).getX());
-            enclosing.setYMin(shape.getPoints().get(0).getY());
-            enclosing.setXMax(shape.getPoints().get(0).getX());
-            enclosing.setYMax(shape.getPoints().get(0).getY());
+            enclosing.setXMin((int) shape.getPoints().get(0).getX());
+            enclosing.setYMin((int) shape.getPoints().get(0).getY());
+            enclosing.setXMax((int) shape.getPoints().get(0).getX());
+            enclosing.setYMax((int) shape.getPoints().get(0).getY());
             for (Point point : shape.getPoints()) {
                 if (point.getX() > enclosing.getXMax()) {
-                    enclosing.setXMax(point.getX());
+                    enclosing.setXMax((int) point.getX());
                 }
                 if (point.getY() > enclosing.getYMax()) {
-                    enclosing.setYMax(point.getY());
+                    enclosing.setYMax((int) point.getY());
                 }
                 if (point.getX() < enclosing.getXMin()) {
-                    enclosing.setXMin(point.getX());
+                    enclosing.setXMin((int) point.getX());
                 }
                 if (point.getY() < enclosing.getYMin()) {
-                    enclosing.setYMin(point.getY());
+                    enclosing.setYMin((int) point.getY());
                 }
             }
         } else {
